refactor(pages): unify hero section names and update images to webp
