+++ src/app/mentors/page.tsx
@@ -1,4 +1,4 @@
-import MentorsHero from "@/sections/mentors/hero/MentorsHero"
+import MentorsHero from "@/sections/mentors/hero/Hero"
 
 export default function PartnersPage() {
   return (
