+++ src/app/partners/page.tsx
@@ -1,4 +1,4 @@
-import PartnersHero from "@/sections/partners/hero/PartnersHero"
+import PartnersHero from "@/sections/partners/hero/Hero"
 
 export default function PartnersPage() {
   return (
