+++ src/sections/collaborators/hero/Hero.tsx
@@ -1,27 +1,25 @@
-import HeroImage from "@/assets/images/photos/hero/image_hero_mentor.png"
+import HeroImage from "@/assets/images/photos/photo_heroCollaborator.webp"
 import HeroLayout from "@/sections/shared/heroLayout/HeroLayout"
 
-export default function MentorsHero() {
+export default function CollaboratorsHero() {
   return (
     <HeroLayout
-      id="mentors"
-      eyebrow="Join SheHub as a mentor"
+      id="collaborators"
+      eyebrow="Join SheHub as a collaborator"
       title={
         <>
-          Support the next generation in tech and {" "}
-          <span
-            className="bg-clip-text text-transparent"
-            style={{ backgroundImage: "var(--color-gradient-brand)" }}
-          >
-            grow as a leader
-          </span>
+          Gain{" "}
+          <span className="bg-clip-text text-transparent" style={{ backgroundImage: "var(--color-gradient-brand)" }}>
+            real tech experience
+          </span>{" "}
+          and grow your portfolio
         </>
       }
       paragraph="At SheHub, you’ll join real tech teams where you can gain hands-on experience, grow with peers, and build your portfolio. Whether you're exploring product management, UX design, development, data, QA or marketing, you’ll work on impactful projects with support from mentors and a community that empowers you."
       mainImage={HeroImage.src}
-      alt="SheHub mentors  collaborating"
+      alt="Collaborators working together"
       buttons={[
-        { text: "Become a mentor", variant: "primary-primary" },
+        { text: "Become a collaborator", variant: "primary-primary" },
       ]}
     />
   )
